refactor: remove the useless crd directory
