+++ spring-cloud-alibaba-starters/spring-cloud-alibaba-commons/src/main/java/com/alibaba/cloud/commons/governance/labelrouting/rule/HeaderRule.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.alibaba.cloud.commons.governance.labelrouting.crd.rule;
+package com.alibaba.cloud.commons.governance.labelrouting.rule;
 
 import java.util.Objects;
