+++ spring-cloud-alibaba-starters/spring-cloud-alibaba-governance/resource-transform/istio-resource-transform/src/main/java/com/alibaba/cloud/governance/istio/protocol/impl/RdsProtocol.java
@@ -22,12 +22,12 @@ import java.util.List;
 import java.util.Map;
 
 import com.alibaba.cloud.commons.governance.event.LabelRoutingDataChangedEvent;
-import com.alibaba.cloud.commons.governance.labelrouting.crd.LabelRouteRule;
-import com.alibaba.cloud.commons.governance.labelrouting.crd.MatchService;
-import com.alibaba.cloud.commons.governance.labelrouting.crd.UnifiedRouteDataStructure;
-import com.alibaba.cloud.commons.governance.labelrouting.crd.rule.HeaderRule;
-import com.alibaba.cloud.commons.governance.labelrouting.crd.rule.RouteRule;
-import com.alibaba.cloud.commons.governance.labelrouting.crd.rule.UrlRule;
+import com.alibaba.cloud.commons.governance.labelrouting.LabelRouteRule;
+import com.alibaba.cloud.commons.governance.labelrouting.MatchService;
+import com.alibaba.cloud.commons.governance.labelrouting.UnifiedRouteDataStructure;
+import com.alibaba.cloud.commons.governance.labelrouting.rule.HeaderRule;
+import com.alibaba.cloud.commons.governance.labelrouting.rule.RouteRule;
+import com.alibaba.cloud.commons.governance.labelrouting.rule.UrlRule;
 import com.alibaba.cloud.commons.lang.StringUtils;
 import com.alibaba.cloud.commons.matcher.StringMatcher;
 import com.alibaba.cloud.commons.matcher.StringMatcherType;
